Add CSGeometry::LoadGeo() interface.
This gives the CSGeometry class the ability to read CSG files via:

CSGeometry geom;
ifstream ist(filename);
CSGeometry * geom_ptr = geom.LoadGeo(ist);

The CSGeometry::Load() interface did not support this (it is
apparently for some other filetype?).  The basic approach is to copy
the code which calls ParseCSG from the CSGeometryRegister class, which
is not accessible in the nglib interface.